refactor: Resolve P2 code review issues
- Split interface file into separate types (GenLauncherDetectionResult, GenLauncherNormalizationResult)
- Use HashSet instead of arrays for O(1) lookup performance
- Reference GenLauncherConstants directly instead of duplicating values
- Add safety check and warning before File.Move overwrite
- Add directory symlink detection
- Use constant for .big extension
